docs/irc-whatwg/2011-11-13.html

<!doctype html>
<head>
  <title>#whatwg on 2011-11-13</title>
  <link rel="stylesheet" href="../style.css">
  <script src="../logs.js"></script>
</head>
<body><div class="wrapper">
<div class="sidebar">
<div class="title">#whatwg<br>2011-11-13<br><a href="plaintext/">plaintext logs</a></div>
<div class="nav">
<a href="2011-11-12" class="nav-link"><span>prev</span></a>
<a href="2011-11-14" class="nav-link"><span style="float:right">next</span></a>
</div>
    
<div class="room-list-wrapper">Active:<br>
<ul class="room-list">
<li><a href="../TC39_Delegates/">TC39 Delegates</a></li>
<li><a href="../TC39_Editors/">TC39 Editors</a></li>
<li><a href="../TC39_General/">TC39 General</a></li>
<li><a href="../WHATWG/">WHATWG</a></li>
</ul>
</div>
<br>
<div class="room-list-wrapper">Historical:<br>
<ul class="room-list">
<li><a href="../irc-tc39/">#tc39</a></li>
<li><a href="../irc-tc39-delegates/">#tc39-delegates</a></li>
<li><a href="../irc-tc39-editor-group/">#tc39-editor-group</a></li>
<li><a href="../irc-whatwg/" class="current-room">#whatwg</a></li>
</ul>
</div>
<div class="footer"><a href="https://github.com/bakkot/matrix-archive-bot">source on github</a></div>
</div>
<div class="rhs">
<div class="rhs-header">
<span id="error" style="color: red; display:none">error</span>
<input type="text" id="query" size=25 placeholder="Search #whatwg">
<a id="search-submit" class="button icon-link" title="Search">
  <svg class="icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path d="M505 442.7L405.3 343c-4.5-4.5-10.6-7-17-7H372c27.6-35.3 44-79.7 44-128C416 93.1 322.9 0 208 0S0 93.1 0 208s93.1 208 208 208c48.3 0 92.7-16.4 128-44v16.3c0 6.4 2.5 12.5 7 17l99.7 99.7c9.4 9.4 24.6 9.4 33.9 0l28.3-28.3c9.4-9.4 9.4-24.6.1-34zM208 336c-70.7 0-128-57.2-128-128 0-70.7 57.2-128 128-128 70.7 0 128 57.2 128 128 0 70.7-57.2 128-128 128z"></path></svg>
</a>
</div>

<div class="log"><table id="log-table"><tbody id="log-tbody">

  <tr class="msg" id="L0"><td class="ts-cell"><a class="ts" href="#L0" alt="Sun Nov 13 2011 09:46:00 GMT-0800 (Pacific Standard Time)">17:46</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">Is there a good way to test [[Configurable]]?</td></tr>
  <tr class="msg" id="L1"><td class="ts-cell"><a class="ts" href="#L1" alt="Sun Nov 13 2011 09:46:01 GMT-0800 (Pacific Standard Time)">17:46</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">Just delete and see if it works?</td></tr>
  <tr class="msg" id="L2"><td class="ts-cell"><a class="ts" href="#L2" alt="Sun Nov 13 2011 09:46:02 GMT-0800 (Pacific Standard Time)">17:46</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=_bga@irc>_bga</span>&gt;</div></td><td class="msg-cell">getOPD?</td></tr>
  <tr class="msg" id="L3"><td class="ts-cell"><a class="ts" href="#L3" alt="Sun Nov 13 2011 09:47:00 GMT-0800 (Pacific Standard Time)">17:47</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">Oh, awesome.</td></tr>
  <tr class="msg" id="L4"><td class="ts-cell"><a class="ts" href="#L4" alt="Sun Nov 13 2011 09:47:01 GMT-0800 (Pacific Standard Time)">17:47</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">Thanks!</td></tr>
  <tr class="msg" id="L5"><td class="ts-cell"><a class="ts" href="#L5" alt="Sun Nov 13 2011 09:48:00 GMT-0800 (Pacific Standard Time)">17:48</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">I guess if that returns the right thing but it doesn&#039;t behave right, that&#039;s more of an ES bug than a WebIDL bug.</td></tr>
  <tr class="msg" id="L6"><td class="ts-cell"><a class="ts" href="#L6" alt="Sun Nov 13 2011 10:01:00 GMT-0800 (Pacific Standard Time)">18:01</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=_bga@irc>_bga</span>&gt;</div></td><td class="msg-cell">AryehGregor <a href="http://pastie.org/2857719">http://pastie.org/2857719</a></td></tr>
  <tr class="msg" id="L7"><td class="ts-cell"><a class="ts" href="#L7" alt="Sun Nov 13 2011 10:05:00 GMT-0800 (Pacific Standard Time)">18:05</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">Only IE9 ever passes my new shiny assert_interface().</td></tr>
  <tr class="msg" id="L8"><td class="ts-cell"><a class="ts" href="#L8" alt="Sun Nov 13 2011 10:05:01 GMT-0800 (Pacific Standard Time)">18:05</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">Go IE.</td></tr>
  <tr class="msg" id="L9"><td class="ts-cell"><a class="ts" href="#L9" alt="Sun Nov 13 2011 10:06:00 GMT-0800 (Pacific Standard Time)">18:06</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">Gecko doesn&#039;t have Node.prototype.constructor.</td></tr>
  <tr class="msg" id="L10"><td class="ts-cell"><a class="ts" href="#L10" alt="Sun Nov 13 2011 10:07:00 GMT-0800 (Pacific Standard Time)">18:07</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">Node.prototype is enumerable and configurable in Opera, and should be neither.</td></tr>
  <tr class="msg" id="L11"><td class="ts-cell"><a class="ts" href="#L11" alt="Sun Nov 13 2011 10:07:01 GMT-0800 (Pacific Standard Time)">18:07</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">In WebKit, window.Node is not writable but is enumerable, while it should be writable but not enumerable.  And String(Node) is wrong.</td></tr>
  <tr class="msg" id="L12"><td class="ts-cell"><a class="ts" href="#L12" alt="Sun Nov 13 2011 10:10:00 GMT-0800 (Pacific Standard Time)">18:10</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">AryehGregor, can you also try it on NodeList?</td></tr>
  <tr class="msg" id="L13"><td class="ts-cell"><a class="ts" href="#L13" alt="Sun Nov 13 2011 10:11:00 GMT-0800 (Pacific Standard Time)">18:11</a></td><td class="nick-cell"><div class="m-ov"><span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span></div></td><td class="msg-cell">tries</td></tr>
  <tr class="msg" id="L14"><td class="ts-cell"><a class="ts" href="#L14" alt="Sun Nov 13 2011 10:12:00 GMT-0800 (Pacific Standard Time)">18:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">Ms2ger, works.</td></tr>
  <tr class="msg" id="L15"><td class="ts-cell"><a class="ts" href="#L15" alt="Sun Nov 13 2011 10:12:01 GMT-0800 (Pacific Standard Time)">18:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">I mean, it passes everything.</td></tr>
  <tr class="msg" id="L16"><td class="ts-cell"><a class="ts" href="#L16" alt="Sun Nov 13 2011 10:12:02 GMT-0800 (Pacific Standard Time)">18:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Excellent</td></tr>
  <tr class="msg" id="L17"><td class="ts-cell"><a class="ts" href="#L17" alt="Sun Nov 13 2011 10:54:00 GMT-0800 (Pacific Standard Time)">18:54</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">What&#039;s an example of a static operation that&#039;s actually used in real life?  I don&#039;t find any in DOM4, and the only one I find in HTML is URL.createObjectURL.</td></tr>
  <tr class="msg" id="L18"><td class="ts-cell"><a class="ts" href="#L18" alt="Sun Nov 13 2011 10:59:00 GMT-0800 (Pacific Standard Time)">18:59</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">IDB</td></tr>
  <tr class="msg" id="L19"><td class="ts-cell"><a class="ts" href="#L19" alt="Sun Nov 13 2011 10:59:01 GMT-0800 (Pacific Standard Time)">18:59</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Those are the only two</td></tr>
  <tr class="msg" id="L20"><td class="ts-cell"><a class="ts" href="#L20" alt="Sun Nov 13 2011 11:03:00 GMT-0800 (Pacific Standard Time)">19:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=AryehGregor@irc>AryehGregor</span>&gt;</div></td><td class="msg-cell">jgraham, I submitted another testharness.js patch: <a href="http://www.w3.org/Bugs/Public/show_bug.cgi?id=14814">http://www.w3.org/Bugs/Public/show_bug.cgi?id=14814</a></td></tr>
  <tr class="msg" id="L21"><td class="ts-cell"><a class="ts" href="#L21" alt="Sun Nov 13 2011 11:30:00 GMT-0800 (Pacific Standard Time)">19:30</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">AryehGregor, also you can test Object.getOwnPropertyDescriptor(obj, propname).configurable</td></tr>
  <tr class="msg" id="L22"><td class="ts-cell"><a class="ts" href="#L22" alt="Sun Nov 13 2011 11:31:00 GMT-0800 (Pacific Standard Time)">19:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">heycam, yes, that&#039;s what _bga said</td></tr>
  <tr class="msg" id="L23"><td class="ts-cell"><a class="ts" href="#L23" alt="Sun Nov 13 2011 11:32:00 GMT-0800 (Pacific Standard Time)">19:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">Ms2ger, I&#039;d have realised that if I&#039;d read just one line further, I suppose :)</td></tr>
  <tr class="msg" id="L24"><td class="ts-cell"><a class="ts" href="#L24" alt="Sun Nov 13 2011 11:32:01 GMT-0800 (Pacific Standard Time)">19:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">:)</td></tr>
  <tr class="msg" id="L25"><td class="ts-cell"><a class="ts" href="#L25" alt="Sun Nov 13 2011 13:37:00 GMT-0800 (Pacific Standard Time)">21:37</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">AryehGregor: getOPD often returns bogus things for host objects in general.</td></tr>
  <tr class="msg" id="L26"><td class="ts-cell"><a class="ts" href="#L26" alt="Sun Nov 13 2011 13:38:00 GMT-0800 (Pacific Standard Time)">21:38</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Well, then that&#039;s an excellent thing to test</td></tr>
  <tr class="msg" id="L27"><td class="ts-cell"><a class="ts" href="#L27" alt="Sun Nov 13 2011 13:38:01 GMT-0800 (Pacific Standard Time)">21:38</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">A WebIDL mapping issue rather than a per-interface issue in general, though</td></tr>
  <tr class="msg" id="L28"><td class="ts-cell"><a class="ts" href="#L28" alt="Sun Nov 13 2011 13:38:02 GMT-0800 (Pacific Standard Time)">21:38</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">And I&#039;m not sure WebIDL actually requires it to be correct</td></tr>
  <tr class="msg" id="L29"><td class="ts-cell"><a class="ts" href="#L29" alt="Sun Nov 13 2011 13:39:00 GMT-0800 (Pacific Standard Time)">21:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">hopefully it does</td></tr>
  <tr class="msg" id="L30"><td class="ts-cell"><a class="ts" href="#L30" alt="Sun Nov 13 2011 13:40:00 GMT-0800 (Pacific Standard Time)">21:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">I have some draft email somewhere about things that WebIDL fails to specify.</td></tr>
  <tr class="msg" id="L31"><td class="ts-cell"><a class="ts" href="#L31" alt="Sun Nov 13 2011 13:40:01 GMT-0800 (Pacific Standard Time)">21:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">I know what your intention is, heycam :)</td></tr>
  <tr class="msg" id="L32"><td class="ts-cell"><a class="ts" href="#L32" alt="Sun Nov 13 2011 13:40:02 GMT-0800 (Pacific Standard Time)">21:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Send it, dammit :)</td></tr>
  <tr class="msg" id="L33"><td class="ts-cell"><a class="ts" href="#L33" alt="Sun Nov 13 2011 13:40:03 GMT-0800 (Pacific Standard Time)">21:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">gsnedders, yes please send :)</td></tr>
  <tr class="msg" id="L34"><td class="ts-cell"><a class="ts" href="#L34" alt="Sun Nov 13 2011 13:43:00 GMT-0800 (Pacific Standard Time)">21:43</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">Oh, no, it&#039;s fixed now anyway</td></tr>
  <tr class="msg" id="L35"><td class="ts-cell"><a class="ts" href="#L35" alt="Sun Nov 13 2011 13:44:00 GMT-0800 (Pacific Standard Time)">21:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">:)</td></tr>
  <tr class="msg" id="L36"><td class="ts-cell"><a class="ts" href="#L36" alt="Sun Nov 13 2011 13:46:00 GMT-0800 (Pacific Standard Time)">21:46</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">yay</td></tr>
  <tr class="msg" id="L37"><td class="ts-cell"><a class="ts" href="#L37" alt="Sun Nov 13 2011 14:21:00 GMT-0800 (Pacific Standard Time)">22:21</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">heh</td></tr>
  <tr class="msg" id="L38"><td class="ts-cell"><a class="ts" href="#L38" alt="Sun Nov 13 2011 14:22:00 GMT-0800 (Pacific Standard Time)">22:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">HTML5 new featured by WSJ: <a href="http://online.wsj.com/article/SB10001424052970203537304577030033160849296.html">http://online.wsj.com/article/SB10001424052970203537304577030033160849296.html</a></td></tr>
  <tr class="msg" id="L39"><td class="ts-cell"><a class="ts" href="#L39" alt="Sun Nov 13 2011 14:22:01 GMT-0800 (Pacific Standard Time)">22:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">&quot;Implementations that use ECMAScript to implement the APIs defined in this specification must implement them in a manner consistent with the ECMAScript Bindings defined in the Web IDL specification&quot;</td></tr>
  <tr class="msg" id="L40"><td class="ts-cell"><a class="ts" href="#L40" alt="Sun Nov 13 2011 14:22:02 GMT-0800 (Pacific Standard Time)">22:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Does that mean C++ implementations are free to do whatever?</td></tr>
  <tr class="msg" id="L41"><td class="ts-cell"><a class="ts" href="#L41" alt="Sun Nov 13 2011 14:22:03 GMT-0800 (Pacific Standard Time)">22:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">I hope so</td></tr>
  <tr class="msg" id="L42"><td class="ts-cell"><a class="ts" href="#L42" alt="Sun Nov 13 2011 14:23:00 GMT-0800 (Pacific Standard Time)">22:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">&quot;implement&quot; there is a bit iffy</td></tr>
  <tr class="msg" id="L43"><td class="ts-cell"><a class="ts" href="#L43" alt="Sun Nov 13 2011 14:23:01 GMT-0800 (Pacific Standard Time)">22:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">it&#039;s not the implementation of the DOM needing to be js</td></tr>
  <tr class="msg" id="L44"><td class="ts-cell"><a class="ts" href="#L44" alt="Sun Nov 13 2011 14:23:02 GMT-0800 (Pacific Standard Time)">22:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">we should definitely not make requirements on other languages I think</td></tr>
  <tr class="msg" id="L45"><td class="ts-cell"><a class="ts" href="#L45" alt="Sun Nov 13 2011 14:23:03 GMT-0800 (Pacific Standard Time)">22:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">it should be saying if the APIs are exposed to JS</td></tr>
  <tr class="msg" id="L46"><td class="ts-cell"><a class="ts" href="#L46" alt="Sun Nov 13 2011 14:23:04 GMT-0800 (Pacific Standard Time)">22:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">or something</td></tr>
  <tr class="msg" id="L47"><td class="ts-cell"><a class="ts" href="#L47" alt="Sun Nov 13 2011 14:24:00 GMT-0800 (Pacific Standard Time)">22:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">heycam: any chance we can get unions in?</td></tr>
  <tr class="msg" id="L48"><td class="ts-cell"><a class="ts" href="#L48" alt="Sun Nov 13 2011 14:24:01 GMT-0800 (Pacific Standard Time)">22:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">heycam: if not, I think I&#039;m just going to use them when I get back to editing DOM4</td></tr>
  <tr class="msg" id="L49"><td class="ts-cell"><a class="ts" href="#L49" alt="Sun Nov 13 2011 14:24:02 GMT-0800 (Pacific Standard Time)">22:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">ha</td></tr>
  <tr class="msg" id="L50"><td class="ts-cell"><a class="ts" href="#L50" alt="Sun Nov 13 2011 14:24:03 GMT-0800 (Pacific Standard Time)">22:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">heycam: much like Hixie&#039;s has been using non-existing features</td></tr>
  <tr class="msg" id="L51"><td class="ts-cell"><a class="ts" href="#L51" alt="Sun Nov 13 2011 14:24:04 GMT-0800 (Pacific Standard Time)">22:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">I think they make sense</td></tr>
  <tr class="msg" id="L52"><td class="ts-cell"><a class="ts" href="#L52" alt="Sun Nov 13 2011 14:25:00 GMT-0800 (Pacific Standard Time)">22:25</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">dunno if it is worth trying to drop all the overload stuff just in favour of unions</td></tr>
  <tr class="msg" id="L53"><td class="ts-cell"><a class="ts" href="#L53" alt="Sun Nov 13 2011 14:25:01 GMT-0800 (Pacific Standard Time)">22:25</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Btw, cabal, please review the mouse lock spec</td></tr>
  <tr class="msg" id="L54"><td class="ts-cell"><a class="ts" href="#L54" alt="Sun Nov 13 2011 14:25:02 GMT-0800 (Pacific Standard Time)">22:25</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">might be possible</td></tr>
  <tr class="msg" id="L55"><td class="ts-cell"><a class="ts" href="#L55" alt="Sun Nov 13 2011 14:25:03 GMT-0800 (Pacific Standard Time)">22:25</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Ms2ger: place a pointer in</td></tr>
  <tr class="msg" id="L56"><td class="ts-cell"><a class="ts" href="#L56" alt="Sun Nov 13 2011 14:25:04 GMT-0800 (Pacific Standard Time)">22:25</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">though I&#039;m going to try to sleep shortly</td></tr>
  <tr class="msg" id="L57"><td class="ts-cell"><a class="ts" href="#L57" alt="Sun Nov 13 2011 14:25:05 GMT-0800 (Pacific Standard Time)">22:25</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell"><a href="http://dvcs.w3.org/hg/webevents/raw-file/default/mouse-lock.html">http://dvcs.w3.org/hg/webevents/raw-file/default/mouse-lock.html</a></td></tr>
  <tr class="msg" id="L58"><td class="ts-cell"><a class="ts" href="#L58" alt="Sun Nov 13 2011 14:25:06 GMT-0800 (Pacific Standard Time)">22:25</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">still adjusting after the flight :(</td></tr>
  <tr class="msg" id="L59"><td class="ts-cell"><a class="ts" href="#L59" alt="Sun Nov 13 2011 14:26:00 GMT-0800 (Pacific Standard Time)">22:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">heycam: there&#039;s definitely some overlap</td></tr>
  <tr class="msg" id="L60"><td class="ts-cell"><a class="ts" href="#L60" alt="Sun Nov 13 2011 14:26:01 GMT-0800 (Pacific Standard Time)">22:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Hmm, sleeping...</td></tr>
  <tr class="msg" id="L61"><td class="ts-cell"><a class="ts" href="#L61" alt="Sun Nov 13 2011 14:26:02 GMT-0800 (Pacific Standard Time)">22:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">heycam: though the append() stuff you cannot really do with overloading because of the varargs</td></tr>
  <tr class="msg" id="L62"><td class="ts-cell"><a class="ts" href="#L62" alt="Sun Nov 13 2011 14:27:00 GMT-0800 (Pacific Standard Time)">22:27</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">allowing append(Comment, Text) does not work with append(Comment...); append(Text...) afaict</td></tr>
  <tr class="msg" id="L63"><td class="ts-cell"><a class="ts" href="#L63" alt="Sun Nov 13 2011 14:28:00 GMT-0800 (Pacific Standard Time)">22:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">via this new &quot;Activity&quot; feature of twitter I find out that people &quot;favorite&quot; tweets a lot</td></tr>
  <tr class="msg" id="L64"><td class="ts-cell"><a class="ts" href="#L64" alt="Sun Nov 13 2011 14:28:01 GMT-0800 (Pacific Standard Time)">22:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=heycam@irc>heycam</span>&gt;</div></td><td class="msg-cell">annevk, yeah that wouldn&#039;t allow append(aComment, aComment, aText, aComment)</td></tr>
  <tr class="msg" id="L65"><td class="ts-cell"><a class="ts" href="#L65" alt="Sun Nov 13 2011 15:16:00 GMT-0800 (Pacific Standard Time)">23:16</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">so much for sleeping before midnight :)</td></tr>
  <tr class="msg" id="L66"><td class="ts-cell"><a class="ts" href="#L66" alt="Sun Nov 13 2011 15:17:00 GMT-0800 (Pacific Standard Time)">23:17</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell"><a href="http://imgur.com/K7D7Y">http://imgur.com/K7D7Y</a> hehe</td></tr>
  <tr class="msg" id="L67"><td class="ts-cell"><a class="ts" href="#L67" alt="Sun Nov 13 2011 15:17:01 GMT-0800 (Pacific Standard Time)">23:17</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">jgraham: <a href="http://tokenizer.py:184">tokenizer.py:184</a>(entitiesStartingWith) is a huge bottleneck in html5lib/PyPy, accounting for a 1/6th of complete.html parse-time.</td></tr>
  <tr class="msg" id="L68"><td class="ts-cell"><a class="ts" href="#L68" alt="Sun Nov 13 2011 15:17:02 GMT-0800 (Pacific Standard Time)">23:17</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">Guess we should move over to using a trie.</td></tr>
  <tr class="msg" id="L69"><td class="ts-cell"><a class="ts" href="#L69" alt="Sun Nov 13 2011 15:17:03 GMT-0800 (Pacific Standard Time)">23:17</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">I mean, we could cache it, but that seems hacky, and will just end up using way more memory than doing it properly.</td></tr>
  <tr class="msg" id="L70"><td class="ts-cell"><a class="ts" href="#L70" alt="Sun Nov 13 2011 15:19:00 GMT-0800 (Pacific Standard Time)">23:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=Philip`@irc>Philip`</span>&gt;</div></td><td class="msg-cell">Alternatively, hardcode the entities lt/gt/amp/quot because they&#039;re a vast majority of all occurrences</td></tr>
  <tr class="msg" id="L71"><td class="ts-cell"><a class="ts" href="#L71" alt="Sun Nov 13 2011 15:19:01 GMT-0800 (Pacific Standard Time)">23:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=Philip`@irc>Philip`</span>&gt;</div></td><td class="msg-cell">(and fall back to a slower search if it&#039;s none of them)</td></tr>
  <tr class="msg" id="L72"><td class="ts-cell"><a class="ts" href="#L72" alt="Sun Nov 13 2011 15:20:00 GMT-0800 (Pacific Standard Time)">23:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=Philip`@irc>Philip`</span>&gt;</div></td><td class="msg-cell">(Also, look at data to verify the hypothesis that those entities are the vast majority)</td></tr>
  <tr class="msg" id="L73"><td class="ts-cell"><a class="ts" href="#L73" alt="Sun Nov 13 2011 15:29:00 GMT-0800 (Pacific Standard Time)">23:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">Philip`: Doesn&#039;t help seeming it&#039;s scanned one-character-at-a-time</td></tr>
  <tr class="msg" id="L74"><td class="ts-cell"><a class="ts" href="#L74" alt="Sun Nov 13 2011 15:32:00 GMT-0800 (Pacific Standard Time)">23:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=Philip`@irc>Philip`</span>&gt;</div></td><td class="msg-cell">See if first char is l/g/a/q; if so, see if <a href="http://self.stream">self.stream</a>.char() is t/t/m/u; etc; then when it falls off the fast path, send the whole list of characters to the slower lookup code</td></tr>
  <tr class="msg" id="L75"><td class="ts-cell"><a class="ts" href="#L75" alt="Sun Nov 13 2011 15:33:00 GMT-0800 (Pacific Standard Time)">23:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=Philip`@irc>Philip`</span>&gt;</div></td><td class="msg-cell">Might avoid the cost of doing a load of expensive dictionary lookups (as with a trie) for the common cases</td></tr>
  <tr class="msg" id="L76"><td class="ts-cell"><a class="ts" href="#L76" alt="Sun Nov 13 2011 15:35:00 GMT-0800 (Pacific Standard Time)">23:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">Seems to be the slowest thing in Python 2.7.2 too</td></tr>

</tbody></table></div></div></div></body>
